ResourceLoader: Use more readable cache group for MessageBlobStore
Follows the amespace-ification of ResourceLoader (Id08a220e1d608,
3e2653f8), which implicitly modified the cache keygroup from
"MessageBlobStore" to "MediaWiki_ResourceLoader_MessageBlobStore".

Bring it more in line with other key groups by adopting the same
naming convention as other memcached keys in ResourceLoader and
more generally in MediaWiki code.

Also, factor out the cache key formula so that it is naturally correct
instead of requiring each of the three pairs to be tested and reviewed
to be identical to its counter-part, and yet different from the two
other very similar pairs.

Bug: T308718
Change-Id: I6b0b87e1c5ddf147788b0c61f401e8ff671b843c