SessionManager: write default edit token during session creation
My testing of pywikibot indicates that it's normal for it to request an
edit token from a session that doesn't have one already using a GET
request, and then to POST that edit token very shortly afterwards to
save an edit. This is a stress test of cross-DC session replication. In
my test, replication won and the edit token existed, but there was only
80ms of margin.

If we create the edit token at the earliest opportunity, we should have
at least a few hundred extra milliseconds for replication to occur.

Bug: T279664
Change-Id: I5f60bbe23f6981b0215888fc62b6c2bae68ad8f8