Deprecate the use of magic variables whose names contain a colon
This reduces ambiguity between a parser function invocation (where the
colon separates the first argument) and a magic variable invocation
(where the colon is considered part of the magic variable name).

There shouldn't actually be any of these out in the wild, but it is
safer to deprecate than to assume.

Bug: T236813
Change-Id: I69e4f3b794f22a69efb98f5815df61199d077048