mediawiki.util: Add getTargetFromFragment (and percentDecode)
Allows you to take an arbitrary URL fragment and find the target
element, as you would with querySelector(':target') for the current
fragment.

Useful for finding arbitrary fragments, or if the current URL
was set with history.pushState, in which case :target is not
updated.

Used in DiscussionTools (I4ccc3fd) and MobileFrontend (I48d7346).

Requires the web2017-polyfills to support all our JS environments,
but this is skipped on all modern browsers.

Bug: T315872
Change-Id: I605a24c708e60ccffde9dce524485f61fd5c7b06