SpecialRecentChangesLinked: Use rdbms code for building the main query
Currently, the main query is basically built by gluing string which can
easily break in other RDBMS systems.

In order to fix this, I basically wrapped this around a select and made
the main query a subquery in FROM statement.

Also while we are at it, let's add support for max execution time.

Change-Id: I1320d03739968dac446065df747c3ecb99e292d6