rdbms: Use JSON to store ChronologyProtector position data
PHP serialization is brittle, so serialize DBPrimaryPos objects to JSON
structures before writing them to the stash.

BREAKING CHANGE: This adds to methods to the DBPrimaryPos interface.
There are no known implementations of this interface outside MediaWiki
core. There appears to be no good alternative to adding these methods to
the interface. The breaking change was announced on wikitech-l on August 31.

Bug: T316601
Change-Id: I78096d893a6a2c000c35673b78e36b52369560e5