Load potential current ip block from correct wiki
When placing an autoblock we check wether the ip is already blocked by a
user or through another autoblock. In the second case we just reset the
expiry. In the first case we simply do nothing.

The problem here is that DatabaseBlock::newFromTarget(), which is used
here, is not cross-wiki compatible, so the correct blocks will not be
loaded here if the parent block is a cross-wiki block.

So, unfortunately, we have to manually implement the query here so that
the correct blocks are loaded.

Bug: T281972
Bug: T284873
Change-Id: Ia90e3578f7c103c40e88f3b68ddcba4a13734d80