ResourceLoader: Remove deprecated global 'loadedScripts' property
This introduced in MediaWiki 1.12.2 as part of importScriptURI, but
never documented as public API. It was soft-deprecated in 1.17 (2011)
and hard-deprecated in MW 1.25 (2015).

According to Global Search, there were 21 mentions of a identifier
with this name. [1] Most are false positives (e.g. standalone code
that has another property by that name, or mentioned in comments
or regexes only). There were a handful of hacks that try to undo the
tracking so that the same URL can be loaded a second time. I've
converted those to use mw.loader.load() instead. There are no live
references left anywhere on WMF wikis.

[1] https://global-search.toolforge.org/?namespaces=2%2C4%2C8&q=loadedScripts&title=%28Gadgets-definition%7C.%2A%5C.%28js%7Ccss%7Cjson%29%29&purge=1

Change-Id: Ia0356f6d904c0de47843c719ed8c1a15d1c35a00