rdbms: Log warning if ChronologyProtector finds data under presumed clientID
We'd like to remove the unconditional ChronologyProtector roundtrip
form the majority of pageviews and other web requests. This should
only be needed in the few seconds following a db-write action.

Before we remove it, use this logging to confirm our theory that this
never happens in practice, or that it happens only in cases where it
is not needed.

Bug: T314434
Change-Id: I7cc280261eaac14c770dc1c03321c2a7e155db19