IndexPager: Fix paging with setIncludeOffset and multiple index fields
When using `setIncludeOffset( true )`, queries generated by the pager
use `>=` or '<=' for the index field instead of `>` or `<`, in order
to include the specified offset in the results.

However, when the index covers multiple fields (indicated by
`getIndexField()` returning an array or arrays), this must be only
done for the last field. Otherwise the conditions become redundant,
and effectively only the first field of the index is compared.

In short:
  (A, B) >= (1, 2)
is equivalent to:
  ((A > 1) OR (A = 1 AND B >= 2))
and not:
  ((A >= 1) OR (A = 1 AND B >= 2))

Change-Id: I44f975cdd7b8bdd243b6d5968ee0b8d1cb953292