FileCache: Allow ?uselang= to utilize cache if set to content language
Previously the mere presence of uselang made it disable the FileCache,
even though requesting the wiki content language is equivalent
to not providing it at all.

The requested language being the content one is already checked
in lines 124-138.

This used to work but was broken with the 2008 refactoring of
isFileCacheable() in e29274ca (r41274).

Change-Id: I01320d36a3b31b7a69025d845b61d9eb05e9fca3