Refresh some http://*.wikipedia.org URLs
Update the protocol to HTTPS. Update the titles of the Wikipedia
articles mentioned in qqq.json – [[Exchangeable image file format]]
became [[Exif]] and [[Dilution of precision (GPS)]] became [[Dilution of
precision (navigation)]]. Fix the importImages.php --source-wiki-url
documentation – the private methods getFileCommentFromSourceWiki() and
getFileUserFromSourceWiki() directly append /api.php to the URL, so it
needs to include the script path. And update the broken mailing list
archive URLs (manually found the matching threads in the archives for
the respective months).

Change-Id: I1f085532ab720265acb1213f33869f759a987d92