rdbms: Remove unused channelsMatch() from internal MySQLPrimaryPos
Remove it from the interface as well. The interface is marked stable
to implement for (theoretical) database extensions.

This is not a breaking change as any existing implementations of the
method may continue to exist, and continue to be uncalled by core.

The interface is not stable to type or call, and the built-in
implementation is internal/unexposed.

Change-Id: Iced3ab3dfca70dec2958eb9d750e7025b95b7de3