Fix selser on html endpoints
Introduced in If97a74ce251f281b7d980928a01b764d6ec0d0a4

Private wikis don't have a RESTBase to pass in the stashed oldHTML but
still want to do selser.  Parsoid can reparse the provided wikitext for
an oldHTML.  Caveat is that that only works for inlined data attributes,
the pagebundle endpoints don't work, see T146187.

The code here could check that 'from' is 'html' and not 'pagebundle' but
leave that for a follow up with tests.

In the future, we can maybe remove the fallback logic in Parsoid and
have that reparse request happen somewhere "higher up".

Bug: T317215
Change-Id: I7af3680a6775d25e3ea7e6b332dd751202d181c5