Restore compatibility with overrides for IndexPager::makeLink()
Extensions like CheckUser expect IndexPager::makeLink() to be used for
displaying the navigation links of the pager, and it was marked as
'@stable to override'.

If IndexPager::makeLink() is overridden, it will be called instead of
PagerNavigationBuilder::makeLink(). This behavior is considered
deprecated.

Also mark PagerNavigationBuilder::makeLink() as protected and
'@stable to override', to provide a migration path. Users can now
override IndexPager::getNavigationBuilder() to return a subclass
of PagerNavigationBuilder, and override makeLink() on that subclass.

Follow-Up: cfd6ffe7bbadf8177da34460930ca4780bd38524
Bug: T317477
Change-Id: I4e2f36b543462aa5d852733da650fb70d49ebf06