BackLinkCache::queryLinks: Make janky filter more robust
The conds part of the query, could potentially be a string instead
of an array. The individual keys can have values that are not strings
like numbers or null. This casts everything appropriately to be
more robust.

Additionally, this fixes a deprecation warning on php8.1 about
passing null to preg_match.

Bug: T313663
Change-Id: I5671c7882af57be29299e9032c6e8f056fe10e30
(cherry picked from commit 466fa62321db7639627750161959a0858be2513c)