Permissions: Exit early from RestrictionStore::loadRestrictions if no rev id
When a new page is created, MW checks if the page is "protected"
from autopatrolling in the middle of its creation
(WikiPage::doUserEditContent). At this point, there is no latest rev,
but $page->getId() still returns something.

This causes a warning in php8, along with some caching behaviour that's
almost certainly wrong. Exit early when this happens.

Bug: T313663
Change-Id: Iebaa536b1281588bf3db12d724f3d17f3d3b20e1