Add support of ParserOptions::setExternalLinkTarget() in legacy parser tests
As a follow-up to Ife14d7319ecbb7b588cbac97fae109b14230c44a, we want to
check that the legacy parser and Parsoid align on what's expected for
nofollow/noreferrer/noopener. Since the wgExternalLinkTarget setting is
modified on the Parsoid side during specific tests, we align the test
runner on the legacy parser to be able to test for this.

Bug: T186241
Change-Id: I2824182eaa5c50cd3e8e0193726d896da285a3a8