Remove meaningless "..." tooltip from jQuery spinner module
I had to dig really deep to find when these three dots have been
added, and why. They are from 2007:
https://phabricator.wikimedia.org/rMWbf2a2cd7

From all I know about web accessibility this is not helpful for
anybody:
* Screenreaders that are set up to read tooltips will either ignore
  this or read something like "dot dot dot". This doesn't tell the
  user anything.
* Users that are able to see the tooltip learn nothing that would
  not be already obvious from the spinner graphic.
* There is usually not much time to hover the spinner image anyway.
* It would make sense as an alt="…" attribute. And indeed, it was
  introduced as an alt="…" attribute in 2007. But it's not an <img>
  tag any more.
* The duplication as a title never made sense. Still it was dragged
  around for 15 years.

Adding an aria-label="…" might be an option but can be done any time
later. Probably not worth anybodies time because screenreader users
won't be able to access this anyway in the short time it's active.

Bug: T317858
Change-Id: I76ed7dbcfbf2444cd5dd9b6761f902fb3aa39a95