Use more permissive match for TOC_PLACEHOLDER in parser output
Extensions can modify the HTML in various ways which could break a
too-specific search-and-replace for the TOC_PLACEHOLDER.  In particular,
DiscussionTools was postprocessing parser HTML in a way which was
breaking the TOC replacement in Parser::replaceTableOfContentsMarker()
(T317857), but in the future Parsoid might also emit slightly different
TOC markers, eg with additional attributes.

Make the search-and-replace more robust, at perhaps a small performance
cost.

Bug: T317857
Change-Id: Id0065d81bbfbe1bf6bea6af1de39ea7e9d6598d9