rdbms: Limit CP warning message for T314434 to initial 10s window
Follows-up I7cc280261eaa. The data in Logstash shows some requests
we expect (e.g. bots and login.wikimedia.org), but also a high volume
of regular index.php requests from pageviews in the minute after a
DB write.

This is due to us storing the data in Memcached for 60s but only
needing it for 10s (per the cookie). For now, focus the warning
on cases where we know we're within the 10s window where a cookie
could have persisted. If it's outside that range, then it's expected
that we have no clientId coming in and don't need to apply CP.

Bug: T314434
Change-Id: I56900622de9dac1738ce249fd3586d6e6bbb3338