SpecialExpandTemplates: Always use content language when no title is given
When no title is given and fallback to using the special page, the page
language is $wgLang which can be StubUserLang and against some type hints.

Using content language is introduced in commit ec30b2e, but I missed
the other ParserOption below for parsing. Preprocess options would not
affect parsing, so we can simply reuse the previous one.

Bug: T317896
Change-Id: I03dc18409815b2cbe382a68ac7b30982f34fc43e