Change subpages element from <span> to <div>
The subpages element (which produces breadcrumbs on subpages)
is expected to be on a newline on virtually all known mediawiki skins.

This behaviour can be achieved by changing the element to a <div>
instead of adding the extra `display:block` rule in
resources/src/mediawiki.skinning/interface-subtitle.less.

(After the HTML cache for this change has been cleared,
that line of CSS can be removed).

Bug: T317452
Change-Id: If7d3ef5528c934fe57b0c55f1d8f138d40228db4