Have Special:UserGroupRights display user group technical name
Old behavior is to display the human readable name (e.g. Autopatrolled) and a "list of members" link. New behavior is the same, except it also displays the technical name (e.g. autoreviewer) enclosed in <code> tags.

This is in a similar style to the rights that are displayed in the adjacent column.

Bug: T318326
Change-Id: Ic288f6f147745c05005b9392126337d804834532