ApiSandbox: use formatversion=latest by default
Developers using ApiSandbox often don't realize they should be using
the "modern" formatversion=2 because our UX defaults to
formatversion=1.  We can't change the default in the action API due to
b/c, but we can hack things so that when you use ApiSandbox we
initialize the query to set formatversion=latest.  Hopefully this will
encourage adoption of the most modern formatversion and avoid client
bugs caused by the "falsey but present" values returned by
formatversion=1.

Change-Id: Ia7c8582a2589f971f1fb22fdc5afb3ee9bde1d50