objectcache: Fix lack of DI for MultiWriteBagOStuff sub caches
I noticed that things like 'logger', 'keyspace', and 'stats' are
unset in WMF production for the SqlBag embedded in MultiWriteBag.

This is a regression from switching its constructor from
calling ObjectCache::newFromParams to calling ObjectFactory directly.

It happens to have little to no impact on prod, but only thanks to
various coincidences, each of which can change and start to cause
problems.

Bug: T318272
Change-Id: Ia893ddb36427eb5e9bff0a3776a6856d10d01adc