RecentChanges: Mark onRecentChangesPurgeRows() as void
The return value isn't used by RecentChangesUpdateJob, and I can't
think of a use case for one extension to cancel another extensions'
ability to react to this event.

There is currently one extension (ORES) using this hook in
Wikimedia Gerrit, and it does not return a value, and does not
use the RecentChangesPurgeRowsHook interface.

Change-Id: I05fa2ed24208dbea4ad299f6c85b54ee224e4880