exception: Tolerate no service container when trying DB rollback
If the exception handler gets triggered before a service container is
available, it should not attempt a DB rollback. Attempting to access
the DBLoadBalancerFactory service object would cause a secondary error,
obscuring the original issue.

This might happen when the exception handler is triggered during
bootstrapping or in the installer.

Change-Id: I644bd0953aa5e690fea16d9fc11ca3f24cb3f104