Soft-deprecate the remainder of ActorMigration
Since MediaWiki 1.39 (I66b2cb8653), the actor migration stage is
always SCHEMA_COMPAT_NEW, so ActorMigration doesn't really do
anything useful and can be replaced with a hard-coded query.
This patch marks the class as deprecated and adds information
on how to replace it, with the intent of getting backported to
1.39.

See https://www.mediawiki.org/wiki/Actor_migration for background.
Note that this process happened in MediaWiki 1.34 for wikis using
update.php, but the migration code has been kept around so far
because the migration process took a long time for some large
wikis.

Usage check for upcoming hard deprecation:
https://codesearch.wmcloud.org/search/?q=%5CbActorMigration%5Cb&i=nope&files=&excludeFiles=&repos=
- getWhere('rev_user'), getJoin('rev_user'): used in many extensions
  and some core classes
- getJoin('rc_user'): in SocialProfile, already hard-deprecated
- getJoin('log_comment'), getJoin('log_user'): used in
  ArticleFeedbackv5, already hard-deprecated
- getInsertValues('log_user'): used in ArticleFeedbackv5, already
  hard-deprecated
- MIGRATION_STAGE_SCHEMA_COMPAT, class existence check: used in
  UserMerge, DynamicPageList, UniversalLanguageSelector,
  StubUserWikiAuth. Might be used for B/C with old MediaWiki
  versions, there isn't really a proper deprecation path for this
  use case, but it's only relevant for 1.33 and older, so probably
  can be let go now.

Bug: T227047
Change-Id: Ib02e38d20c5a1f96e68a4f67bbd981375c2e97db