html2wt: fall back to re-rendering if needed.
If we don't have a render id given, but we do have a revision id, we can
fall back to the current rendering that revision to provide a baseline for
selser. This is better than no selser. On wikis that do not heavily rely
on templates, or where templates rarely change while an edit is in
progress, this will produce a clean diff.

Bug: T318398
Change-Id: If7612cc6e64d1f1243289b7d6ba96c71f09fe15d