Add Parser::msg() helper for messages from extensions or parser functions
Using a standard wfMessage(...)->escaped() in the context of a parser
function or tag hook is wrong, because it'll use the current user's
language and pollute the cache. In reviewing MediaWiki extensions a
lot of them do this wrong.

This provides a helper to make it more likely that folks will get this
right.

Bug: T202481
Change-Id: Ic6dfd111c173b97a959c27e18fa83cc572326fa6