content: Stop encoding `&` in JavaScript redirect content
There is no security benefit from encoding &, it's perfectly safe in
JavaScript (it likely dates from XML/XHTML requirements).

Newly created redirects will use a literal & in these URLs, while
continuing to support use of \u0026 for existing pages.

Note that this is about use of & for query parameter seperators, the
& in a page title will continue to be encoded as %26 in the 'title'
value and is unaffected by this change.

Bug: T107289
Co-Authored-By: Ammar Abdulhamid <ammarpad@yahoo.com>
Change-Id: I1db4483db6bc52a96487fefd2c3693b4825ccbb2