Parser: Section offsets are in codepoints, not in bytes
The offset is actually mesured in codepoints, not in bytes.

This field is meant to replace the "byteoffset" field since the
current naming is misleading, and we already have misused it in
deployed extensions.  Support was added in Parsoid in
Ide436dca5a609c866da3c63049723243b8242c34 and the patch depends
on a version of Parsoid with that patch in mediawiki-vendor.

Parsoid still uses the old name in the ::toLegacy() serialization, and
thus in the action API, but that method will eventually be deprecated
(T327439, T330232).

Bug: T319141
Depends-On: Iacdd9a11b79bbafb9cfe9568c889ed721a137833
Depends-On: Ide436dca5a609c866da3c63049723243b8242c34
Change-Id: Ie618a964574780d2ad72192483b399407c7a0bbe