jquery.tablesorter: Remove 'url' parser for transforming sorted text
This was a light transformation that just strips the protocol and
then sorts it as text. This will fallback to sorting the URLs directly
as text instead. This means e.g. ftp://a, ftp://b, http://a, http://b
will now group by protocol, and then by domain; instead of previously
where it grouped by domain.

This functionality was not covered by tests.

Bug: T47161
Change-Id: I526cafdc0973d550ce1810e07a10fb39e25e9142