rdbms: remove DB domain parameter from various lag/read-only methods
This effects:
* LoadBalancer::getReadOnlyReason()
* LoadBalancer::getLaggedReplicaMode()
* LoadBalancer::isPrimaryRunningReadOnly()
* LoadBalancer::getRandomNonLagged()

The lag/read-only mode is designed to be tracked per-server. The DB domain
is not useful here, and passing one complicates callers, causing needless
cache fragmentation, and cause can errors when a foreign LoadBalancer falls
back (or is provided) a local DB domain (which might not reside on the server).

Clarified some related "server" vs "DB" terminology in comments.

Bug: T318904
Change-Id: I3e09c1915d012db36be1bc505e00dcb0603fd388