In the event of preg failure in MagicWordArray throw exception
Previously we just logged this. If this happens, it replaces the
entire page with null, which I think is exception worthy and
very confusing if it happens.

The most likely cause of this error would be if the input
text had invalid UTF-8 in it.

I checked logstash, and this warning did not appear in the logs
in the last 90 days.

See also bug T319218

Change-Id: Ic0c9083bd8c524ddce9c7d9d10629daa8f6b8999
(cherry picked from commit f27945728f1dcbc296ff397de68123fe78eec06c)