ApiSandbox: Replace obsolete IntegerParamWidget class
It appears like there was once a reason to do this. But the OOUI
NumberInputWidget changed quite a bit since then. Most notably there
is no "isInteger" option any more. Which is why I started having a
closer look at this file. I believe we can as well inline it, as
suggested in this patch. Note the same is already done for e.g. the
"string", "user", and "expiry" types above.

Change-Id: Ia64ca06f980ac5d40c6f4af500e0b1214bd9f1c5