tests: Add temporary forward-compat method for assertMatchesRegularExpression
assertRegExp was deprecated in PHPUnit 9 and its replacement,
assertMatchesRegularExpression, was also introduced in PHPUnit 9. Hence,
we cannot migrate callers before the upgrade, but we also can't upgrade
before callers are migrated due to deprecation warnings. This is
especially true in extensions.

So create a temporary method that extensions can use while we
transition.

Bug: T243600
Change-Id: Ibce4c2479a533d132eb9f0641f8473fc6ca14fdb