Make IDE not complain on `createNoOpMock()` & `createNoOpAbstractMock()`
Since `createNoOpMock()` & `createNoOpAbstractMock()` returns a `MockObject`
type, this makes callers of this method complain in the IDE as if the right
type is not passed in.

Using a Psalm template in this patch fixes this warnings (got the idea
from PHPUnit TestCase.php). In addition, using normal @params so that
phan can be happy as well.

Change-Id: Ic0074d0d0d432c71473a48d79653e59ca7ad0085