Make LanguageFactory::getParentLanguage() consistent with ::getLanguage()
Allow Bcp47Code (which includes Language) as a parameter to
::getParentLanguage(), to match what was added in
Iea158fc849964ba71a27363f9522b1b3bbb82aee. Add a note to warn
callers that ::getParentLanguage() does *not* accept the same set
of string codes as ::getLanguage(); however if you pass an object
instead of a string the behavior matches.

Change-Id: I89c581ea12b354d8988c1ae8787a05976b8e3143