jquery.tablesorter: Convert custom cases to native QUnit.test()
It seems generally preferred for static analysis benefit, and to
ease debugging and re-running of tests (e.g. QUnit.test.skip,
QUnit.test.only) and for ease of modification (since some already
don't use this shortcut in the same file) that abstractions exist
as utility functions called within the test function rather than
for the utility to register the test itself.

The relevant utility functions to make that easy, already existed
but I guess we never finished the conversion. When finished, it
becomes clear that the previous state did not save lines of code.

Note that in cases where utility functions act as emulated data
providers to generate multiple test cases (not in this file) there
also exists `QUnit.test.each()` to do this natively.

Change-Id: I52d82578357239fdfc5d01cabb09ea6c10d8c397