In Language::ucfirst(), use title case instead of upper case
ucfirst() is an implementation of what Unicode calls title case.
Implementing it by doing mb_strtoupper() on the first character produces
a number of artifacts on PHP 7.3+, most notably destruction of all
titles in the Georgian language.

MB_CASE_TITLE was introduced in PHP 7.3. It does the same thing as
mb_strtoupper() for most characters, but avoids inappropriate mappings
of Georgian characters and ligatures.

It's possible to use $wgOverrideUcfirstCharacters to restore the
original behaviour. This may be useful during migration.

Bug: T292552
Change-Id: If02247fd15ad2a392a488d8162df3e7f63e31d3f